FinalProject/requirements.txt
Flask==1.1.1
WTForms==2.2.1
Flask-WTF==0.14.2
google-cloud-datastore==1.9.0
cerberus==1.3.2
requests-oauthlib==1.3.0
requests==2.22.0
google-cloud-logging==1.12.1

